Relocated getBrandedImage utility
