+++ src/app/lib/utilities/getUrlHelpers/index.js
@@ -1,4 +1,4 @@
-const getBrandedImage = (locator, service) =>
+export const getBrandedImage = (locator, service) =>
   `${process.env.SIMORGH_ICHEF_BASE_URL}/news/1024/branded_${service}/${locator}`;
 
-export default getBrandedImage;
+
